requirements.txt
fastapi==0.103.2
uvicorn==0.32.0
pydantic==1.10.18
sqlalchemy==2.0.36
aiohttp==3.10.10
websockets==9.1
web3==5.31.3
supabase==1.0.3
protobuf==3.19.5
python-dotenv==1.0.1
python-multipart==0.0.12
requests==2.32.3
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
transformers==4.45.2
huggingface-hub==0.25.2
tokenizers==0.20.1
openai==1.51.2
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.1.1
fastapi-cache2==0.2.2
fastapi-limiter==0.1.6